schema: store plaintext messages
basically, i was completely wrong about the former architecture.
it was trivial to get the content of every message. i did that in about 5 minutes.
so basically, message hashing was a waste of CPU time and get_message was a waste of API rate limits.

storing the message content is still not terrible because we don't know who sent it.
i will store message ID, which is enough to handle deletes and edits, but not enough
to handle channel deletions and guild leaves.
In the future, it may turn out to be worth trading off the anonymity of channel info
for the ability to handle channel deletes and such.