Improve shout detection (aka Unicode is pain)
So according to python, 🅱 is uppercase. However, as an emoji, it needs a
"Variation Selector 16" to follow it, like so: 🅱️.
The problem is, variation selector 16 does not count as upppercase,
so a string consisting solely of B button emojis would not count as a shout.

To do this, we ignore the set of "Default Ignorable" code points, as pointed out by
@FakeUnicode on Twitter. <https://twitter.com/FakeUnicode/status/1011855975517548544>.

Note that merely importing is_shout constitutes an increase of 5MiB of RAM usage.
Had we not done `del DERIVED_CORE_PROPERTIES` at the end of the file, is_shout would take up
95MiB of RAM.